backend/requirements.txt
flask
flask-cors
gunicorn
psycopg2
PyJWT
python-dotenv